+++ selenium/src/main/java/software/xdev/tci/selenium/testbase/SeleniumRecordingExtension.java
@@ -16,23 +16,25 @@
 package software.xdev.tci.selenium.testbase;
 
 import java.util.Objects;
+import java.util.Optional;
 import java.util.function.Function;
-import java.util.regex.Pattern;
 
 import org.junit.jupiter.api.extension.AfterTestExecutionCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.testcontainers.lifecycle.TestDescription;
 
+import software.xdev.tci.junit.jupiter.FileSystemFriendlyName;
 import software.xdev.tci.selenium.BrowserTCI;
 
 
+/**
+ * @deprecated Use {@link SeleniumRecorder} instead
+ */
+@Deprecated(forRemoval = true)
 public abstract class SeleniumRecordingExtension implements AfterTestExecutionCallback
 {
 	private static final Logger LOG = LoggerFactory.getLogger(SeleniumRecordingExtension.class);
-	
-	protected static final Pattern FILE_NAME_ALLOWED_CHARS = Pattern.compile("[^A-Za-z0-9#_-]");
 	
 	protected final Function<ExtensionContext, BrowserTCI> tciExtractor;
 	
@@ -43,45 +45,11 @@ public abstract class SeleniumRecordingExtension implements AfterTestExecutionCa
 	}
 	
 	@Override
-	public void afterTestExecution(final ExtensionContext context) throws Exception
+	public void afterTestExecution(final ExtensionContext context)
 	{
-		final BrowserTCI browserTCI = this.tciExtractor.apply(context);
-		if(browserTCI != null)
-		{
-			LOG.debug("Trying to capture video");
-			
-			browserTCI.afterTest(
-				new TestDescription()
-				{
-					@Override
-					public String getTestId()
-					{
-						return this.getFilesystemFriendlyName();
-					}
-					
-					@SuppressWarnings("checkstyle:MagicNumber")
-					@Override
-					public String getFilesystemFriendlyName()
-					{
-						final String testClassName =
-							this.cleanForFilename(context.getRequiredTestClass().getSimpleName());
-						final String displayName = this.cleanForFilename(context.getDisplayName());
-						return System.currentTimeMillis()
-							+ "_"
-							+ testClassName
-							+ "_"
-							// Cut off otherwise file name is too long
-							+ displayName.substring(0, Math.min(displayName.length(), 200));
-					}
-					
-					private String cleanForFilename(final String str)
-					{
-						return FILE_NAME_ALLOWED_CHARS.matcher(str.replace(' ', '_'))
-							.replaceAll("")
-							.toLowerCase();
-					}
-				}, context.getExecutionException());
-		}
-		LOG.debug("AfterTestExecution done");
+		new SeleniumRecorder(LOG).afterTest(
+			context,
+			Optional.ofNullable(this.tciExtractor.apply(context)),
+			new FileSystemFriendlyName(context));
 	}
 }